+++ chempy/tests/test_core.py
@@ -3,4 +3,4 @@ from __future__ import (absolute_import, division, print_function)
 
 
 def test_imports():
-    from .. import MassAction
+    from .. import MassAction  # noqa